Improve exercises instructions
From the README, it is not clear that the user should change into the
`exercises` directory and install the dependencies there. It is
important to mention this because there is indeed a `package.json` in
the `exercises` with a different set of dependencies and scripts to
assert the correctness of the exercises' solutions.