update ch03.md
`slice` and `splice` have different purposes and functionality. I definitely agree that `splice` should be avoided due to its destructive nature. Anyway the only case when they (forgetting for a second about destruction) have the same output is when the first parameter is exactly 0 and optionally both pass the same second integer. Splice